Renamed engine_client.py to external_task_client.py
