+++ camunda/external_task/external_task_worker.py
@@ -3,7 +3,7 @@ import logging
 
 from frozendict import frozendict
 
-from camunda.client.engine_client import ExternalTaskClient
+from camunda.client.external_task_client import ExternalTaskClient
 from camunda.external_task.external_task import ExternalTask
 from camunda.external_task.external_task_executor import ExternalTaskExecutor
 from camunda.utils.log_utils import log_with_context
